+++ src/main/java/net/mohron/skyclaims/integration/nucleus/NucleusIntegration.java
@@ -25,7 +25,6 @@ import io.github.nucleuspowered.nucleus.api.service.NucleusHomeService;
 import io.github.nucleuspowered.nucleus.api.service.NucleusKitService;
 import net.mohron.skyclaims.PluginInfo;
 import net.mohron.skyclaims.SkyClaims;
-import net.mohron.skyclaims.config.type.integration.NucleusConfig;
 import net.mohron.skyclaims.integration.Version;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.entity.living.player.Player;
@@ -50,10 +49,10 @@ public class NucleusIntegration {
 
         metaService = NucleusAPI.getMetaService();
         if (Version.of(metaService.semanticVersion()).compareTo(PluginInfo.NUCLEUS_VERSION) >= 0) {
-            PLUGIN.getLogger().info(String.format("Successfully integrated with Nucleus %s!", metaService.version()));
+            PLUGIN.getLogger().info("Successfully integrated with Nucleus {}!", metaService.version());
         } else {
-            PLUGIN.getLogger().info(String.format("Found Nucleus %s. SkyClaims requires Nucleus %s+... disabling integration",
-                metaService.semanticVersion(), PluginInfo.NUCLEUS_VERSION));
+            PLUGIN.getLogger().info("Found Nucleus {}. SkyClaims requires Nucleus {}+... disabling integration.",
+                metaService.semanticVersion(), PluginInfo.NUCLEUS_VERSION);
             Sponge.getEventManager().unregisterListeners(this);
         }
     }